==> docs/02_Cleaning_DV.Rmd <==
---
title: "02 Cleaning - DV"
author: "J Andres Gannon and Erin Werner"
date: "August 8, 2018"
output:
  html_document:
    theme: flatly
    code_download: true
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(warning = FALSE, error = FALSE)
library(magrittr)
library(ggplot2)
```

# Set-up
## Load packages
Pipe function breaks unless package is loaded. All other packages are referrened only for their specified functions.
```{r}
library(magrittr)
```

## Load and subset data
Load the RData files containing the spreadsheets for the IISS deployment to Afghanistan data and country-year troop levels
```{r}
iiss_deployment <- readRDS(file = paste0(here::here(), '/data/iiss_deployment.rds'))
iiss_troops <- readRDS(file = paste0(here::here(), '/data/iiss_troops.rds'))
```

The country-years in both datasets refer to the IISS publication year which is one year prior to the actual calendar year the data describes
```{r}
iiss_deployment$year <- iiss_deployment$year - 1
iiss_troops$year <- iiss_troops$year - 1
```

Narrow the datasets to just the years 2001-2014, the scope of the Afghanistan war
```{r}
# Subset deployment data
iiss_deployment <- subset(iiss_deployment, year > 2000 & year < 2015)

# Subset troop data
iiss_troops <- subset(iiss_troops, year > 2000 & year < 2015)
```

### Clean names
Clean the country string names from both data frames.
```{r}
iiss_deployment$country[iiss_deployment$country == "Belqium"] <- "Belgium"
iiss_deployment$country[iiss_deployment$country == "Bulqaria"] <- "Bulgaria"
iiss_deployment$country[iiss_deployment$country == "Bosnia -Herzegovina"] <- "Bosnia-Herzegovina" 
iiss_deployment$country[iiss_deployment$country == "Bosnia- Herzegovina"] <- "Bosnia-Herzegovina" 
iiss_deployment$country[iiss_deployment$country == "Bosnia- Herzeqovina"] <- "Bosnia-Herzegovina" 
iiss_deployment$country[iiss_deployment$country == "Bosnia-Herzeqovina"] <- "Bosnia-Herzegovina" 
iiss_deployment$country[iiss_deployment$country == "Burkino Faso"] <- "Burkina Faso"
iiss_deployment$country[iiss_deployment$country == "Hunqary"] <- "Hungary"
iiss_deployment$country[iiss_deployment$country == "Niqeria"] <- "Nigeria"
iiss_deployment$country[iiss_deployment$country == "Portuqal"] <- "Portugal"
iiss_deployment$country[iiss_deployment$country == "Seneqal"] <- "Senegal"
iiss_deployment$country[iiss_deployment$country == "Paraquay"] <- "Paraguay"
iiss_deployment$country[iiss_deployment$country == "Tonqa"] <- "Tonga"
iiss_deployment$country[iiss_deployment$country == "Toqo"] <- "Togo"
iiss_deployment$country[iiss_deployment$country == "Sir Lanka"] <- "Sri Lanka"
iiss_deployment$country[iiss_deployment$country == "China, People's Republic of"] <- "Republic of China"
iiss_deployment$country[iiss_deployment$country == "China, People’ Republic of"] <- "Republic of China"
iiss_deployment$country[iiss_deployment$country == "China, People’s Republic of"] <- "Republic of China"
iiss_deployment$country[iiss_deployment$country == "China, Peoples Republic of"] <- "Republic of China"
iiss_deployment$country[iiss_deployment$country == "Congo, Republic of"] <- "Congo"
iiss_deployment$country[iiss_deployment$country == "Cote d’Ivoire"] <- "Cote D'Ivoire"
iiss_deployment$country[iiss_deployment$country == "Democratic Republic of the Congo"] <- "Democratic Republic of Congo"
iiss_deployment$country[iiss_deployment$country == "Equatorial Guniea"] <- "Equatorial Guinea" 
iiss_deployment$country[iiss_deployment$country == "Guniea"] <- "Guinea"
iiss_deployment$country[iiss_deployment$country == "Kyrgystan"] <- "Kyrgyzstan"
iiss_deployment$country[iiss_deployment$country == "Kyrqyzstan"] <- "Kyrgyzstan"
iiss_deployment$country[iiss_deployment$country == "Macedonia"] <- "Macedonia, Former Yugoslav Republic of"
iiss_deployment$country[iiss_deployment$country == "Macedonia (FYROM)"] <- "Macedonia, Former Yugoslav Republic of"
iiss_deployment$country[iiss_deployment$country == "Macedonia, Former Yugoslav Republic"] <- "Macedonia, Former Yugoslav Republic of"
iiss_deployment$country[iiss_deployment$country == "FYROM"] <- "Macedonia, Former Yugoslav Republic of"
iiss_deployment$country[iiss_deployment$country == "UAE"] <- "United Arab Emirates"
iiss_deployment$country[iiss_deployment$country == "UK"] <- "United Kingdom"
iiss_deployment$country[iiss_deployment$country == "Yemen, Rep of"] <- "Yemen"
iiss_deployment$country[iiss_deployment$country == "Republic of Korea"] <- "Korea, Republic of"
iiss_deployment$country[iiss_deployment$country == "The Gambia"] <- "Gambia"
iiss_deployment$country[iiss_deployment$country == "Timor-Leste"] <- "Timor Leste"
```


# Merge
Merge the two datasets into a single dataframe
```{r,message=FALSE, warning=FALSE, include=FALSE}
years <- c(2001:2014)
myList <- list()
index <- 1

for(yr in years){
  temp_troops <- iiss_troops[which(iiss_troops$year==yr),]
  temp_dep <- iiss_deployment[which(iiss_deployment$year==yr),]
  
  countries <- sort(unique(temp_troops$country))
  
  for(state in countries){
    tt <- temp_troops[which(temp_troops$country==state),]
    dd <- temp_dep[which(temp_dep$country==state),]
    
    active <- tt$troops_active
    
    oefhoa <- dd[which(dd$operational_org=="OEF - HOA"),]
    oh <- as.integer(oefhoa$count_total)
    if(!length(oh)>0){
      oh <- 0
    }
    if(is.na(oh)){
      oh <- 0
    }
    
    isaf <- dd[which(dd$operational_org=="ISAF"),]
    is <- as.integer(isaf$count_total)
    if(!length(is)>0){
      is <- 0
    }
    if(is.na(is)){
      is <- 0
    }
    
    unama <- dd[which(dd$operational_org=="UNAMA"),]
    un <- as.integer(unama$count_total)
    if(!length(un)>0){
      un <- 0
    }
    if(is.na(un)){
      un <- 0
    }
    
    combined <- oh+is+un
    
    if(combined == 0){
      ratio <- 0
    }
    else{
      ratio <- combined/active
    }
     
    myList[[index]] <- data.frame(year = yr, country = state, troops_total = active,
                                  troops_afghan_total = combined, troops_afghan_oef = oh,
                                  troops_afghan_isaf = is, troops_afghan_unama = un,
                                  troops_afghan_ratio = ratio)
    index <- index + 1
  }
}

my_df <- myList[[1]]
for(i in 2:length(myList)){
  my_df <- rbind(my_df,myList[[i]])
}

my_df$ccode <- countrycode::countrycode(my_df$country, 'country.name', 'cown')
my_df$ccode[my_df$country == "Palestinian Territories"] <- "665"
my_df$ccode[my_df$country == "Palestinian Autonomous Areas of Gaza and Jericho"] <- "665"
my_df$ccode[my_df$country == "Republika Sepska Armed Forces"] <- "000"
my_df$ccode[my_df$country == "Serbia"] <- "345"
my_df$ccode[my_df$country == "Serbia and Montenegro"] <- "345"

my_df <- my_df[,c(1,2,9,3,4,5,6,7,8)]

iiss_afghan <- my_df
```

# Save final dataset
```{r}
saveRDS(iiss_afghan, paste0(here::here(),"/data/","iiss_afghan.rds"))
```
