==> docs/03a_SummaryStats_TroopContributions.Rmd <==
---
title: "03a SummaryStats - TroopContributions"
author: "J Andres Gannon, Erin Werner, and Cole Reynolds"
date: "August 8, 2018"
output:
  html_document:
    theme: flatly
    code_download: true
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(warning = FALSE, error = FALSE)
library(magrittr)
library(ggplot2)
```

#Load data
```{r}
troops <- readRDS(file = paste0(here::here(), '/data/iiss_afghan.rds'))
```

# Country contribution ratio histogram
```{r}
troops_hist_largebin(troops)
troops_hist_smallbin(troops)
```


# Annual troop contribution bar plot
```{r}
total_troops_per_year(troops)
total_troops_year_mission(troops)
```

# Troops contribution per country
```{r}
country_troop_lineplot(troops)
```

# Troops contribution distribution boxplot
```{r,warning=FALSE}
country_troop_boxplot(troops)
country_troop_violin_plot(troops)
```

# Troops percentage boxplot
```{r,warning=FALSE}
country_percent_boxplot(troops)
```


# Troops Contribution Maps
## Year 2001 contributions
```{r}
### Filter to only year 2001 observations that contributed troops
troops_2001 <- subset(troops, year == "2001")
troops_2001 <- subset(troops_2001, troops_afghan_total != "0")

### Raw Troop Contributions
map <- rworldmap::joinCountryData2Map(troops_2001,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

rworldmap::mapDevice()

mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='troops_afghan_total',
                                       catMethod = "pretty",
                                       mapTitle = "Troop Contributions to War in Afghanistan (2001)",
                                       addLegend = FALSE
                                      )

do.call(rworldmap::addMapLegendBoxes,
        c(mapParams,
          horiz = FALSE,
          title = "Troop Contribution",
          cex = 0.9
          )
        )

```

```{r}
### Ratio of contribution to total military
map <- rworldmap::joinCountryData2Map(troops_2001,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

rworldmap::mapDevice()

mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='troops_afghan_ratio',
                                       catMethod = "pretty",
                                       mapTitle = "Troop Contributions Relative to Total Military (2001)",
                                       addLegend = FALSE
                                      )

do.call(rworldmap::addMapLegendBoxes,
        c(mapParams,
          horiz = FALSE,
          title = "Contribution Ratio",
          cex = 0.9
          )
        )

```

```{r}
### Now we scale the variables
troops_2001$troops_afghan_total <- scale(troops_2001$troops_afghan_total)
troops_2001$troops_afghan_ratio <- scale(troops_2001$troops_afghan_ratio)
```

```{r}
### Scaled total contribution
map <- rworldmap::joinCountryData2Map(troops_2001,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

rworldmap::mapDevice()

mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='troops_afghan_total',
                                       catMethod = "pretty",
                                       mapTitle = " Total Troop Contributions (2001)",
                                       addLegend = FALSE
                                      )

do.call(rworldmap::addMapLegendBoxes,
        c(mapParams,
          horiz = FALSE,
          title = "Scaled Total Contribution",
          cex = 0.9
          )
        )

```

```{r}
### Scaled Ratio of contribution to total military
map <- rworldmap::joinCountryData2Map(troops_2001,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

rworldmap::mapDevice()

mapParams <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='troops_afghan_ratio',
                                       catMethod = "pretty",
                                       mapTitle = "Troop Contributions Relative to Total Military (2001)",
                                       addLegend = FALSE
                                      )

do.call(rworldmap::addMapLegendBoxes,
        c(mapParams,
          horiz = FALSE,
          title = "Scaled Contribution Ratio",
          cex = 0.9
          )
        )

```

```{r}
library(grid)

### Put plots side-by-side
par(mfrow=c(1,2))
par(mar = c(0,0,0,0))

### Scaled Raw Troop Contributions
map <- rworldmap::joinCountryData2Map(troops_2001,
                                      nameJoinColumn="country",
                                      joinCode="NAME" )

mapParams_1 <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='troops_afghan_total',
                                       catMethod = "pretty",
                                       mapTitle = "",
                                       addLegend = FALSE
                                      )


mapParams_2 <- rworldmap::mapCountryData(map,
                                       nameColumnToPlot='troops_afghan_ratio',
                                       catMethod = "pretty",
                                       mapTitle = "",
                                       addLegend = FALSE
                                      )

grid::grid.text(label = "Total Troops Contributed", x = 0.25, y = 0.73)
grid::grid.text(label = "Percent Troops Contributed", x = 0.75, y = 0.73)
grid::grid.text(label = "Troop Contributions to War in Afghanistan (2001)", x = 0.5, y = 0.85, gp = gpar(fontsize = 14, fontface = "bold"))
grid::grid.text(label = "Darker countries indicate higher values.\nBoth maps scaled relative to average contribution level to make values comparable.", x = 0.5, y = 0.2, gp = gpar(fontsize = 9))


```
