==> docs/03b_SummaryStats_AllianceNetwork.Rmd <==
---
title: "03b SummaryStats - Alliance Network"
author: "Erin Werner and Daniel Kent"
date: "August 8, 2018"
output:
  html_document:
    theme: flatly
    code_download: true
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
start <- read.csv("isaf_start.csv")
alliances <- read.csv("alliance_v4.1_by_dyad_yearly.csv")

isaf_final <- read.csv("polnet_final.csv")


## empty variable for alliance with US or UK
isaf$usuk <- 0

for(i in 1:nrow(isaf)){
  country <- as.integer(isaf[i,2])
  time <- as.integer(isaf[i,3])

  ## select alliances with this country
  temp <- filter(alliances, ccode1 == country | ccode2 == country) %>%
    filter(state_name2 == "United States of America" | state_name2 == "United Kingdom" |
             state_name1 == "United States of America" | state_name1 == "United Kingdom") %>%
    filter(year == time)

  ## any us or uk?
  #temp <- filter(temp, state_name2 == "United States of America" | state_name2 == "United Kingdom")
  ## In this year?
  #temp <- filter(temp, year == time)

  ## Any alliances with US or UK?
  if(nrow(temp) > 0){
    isaf[i,10] <- 1
  }else{
    isaf[i,10] <- 0
  }
}


countries_isaf <- c("Albania",
                    "Armenia",
                    "Australia",
                    "Austria",
                    "Azerbaijan",
                    "Bahrain",
                    "Belgium",
                    "Bosnia and Herzegovina",
                    "Bulgaria",
                    "Canada",
                    "Croatia",
                    "Czech Republic",
                    "Denmark",
                    "El Salvador",
                    "Estonia",
                    "Finland",
                    "France",
                    "Georgia",
                    "Germany",
                    "Greece",
                    "Hungary",
                    "Iceland",
                    #                    "Ireland",
                    "Italy",
                    "Jordan",
                    "Latvia",
                    "Lithuania",
                    "Luxembourg",
                    #                    "Macedonia",
                    "Malaysia",
                    "Mongolia",
                    #                    "Montenegro",
                    "Netherlands",
                    "New Zealand",
                    "Norway",
                    "Poland",
                    "Portugal",
                    "Romania",
                    #                    "Serbia",
                    #                    "Singapore",
                    "Slovakia",
                    #                    "Slovenia",
                    "South Korea",
                    "Spain",
                    "Sweden",
                    #                    "Switzerland",
                    #                    "Tonga",
                    "Turkey",
                    "United Arab Emirates",
                    "United Kingdom",
                    "United States of America",
                    "Ukraine")



ally <- ally[ally$state_name1 %in% countries_isaf &
               ally$state_name2 %in% countries_isaf, ]
ally$alliance_strength <- ally$defense + ally$neutrality + ally$nonaggression + ally$entente
vars_ally <- c("ccode1", "ccode2", "year", "alliance_strength")
ally <- ally[vars_ally]


col_ally <- c("ccode1_name", "ccode2_name")
pair <- ally[col_ally]

values <- c("gray82","darkgray","gray22")
ally$str_color <- values[ally$alliance_strength]

my_igraph <- igraph::graph_from_data_frame(d = pair, vertices = cinc$ccode_name, directed = F)
my_igraph <- igraph::simplify(my_igraph, remove.multiple = T)

deg <- igraph::degree(my_igraph)
igraph::V(my_igraph)$size <- exp(20*cinc$cinc)+8
igraph::V(my_igraph)$shape <- "sphere"
igraph::V(my_igraph)$color <- "cadetblue1"
igraph::E(my_igraph)$width <- (ally$alliance_strength * 0.6)

```
