==> docs/04_DataLoad-Cleaning_EV.Rmd <==
---
title: "04 Data Load and Cleaning - EV"
author: "Daniel Kent"
output:
  html_document:
    theme: flatly
    code_download: true
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(warning = FALSE, error = FALSE)
library(magrittr)
library(ggplot2)
```

# State List

```{R}
states <- readr::read_csv('../inst/extdata/states2016.csv')

isaf_ext <- dplyr::tibble(
    year = as.integer(),
    stateabb = as.character(),
    statenme = as.character(),
    ccode = as.integer()
)

for (i in 1:14){
    year <- 2000 + i ## set year
    ## what countries exist at that time? 
    temp <- dplyr::filter(states, endyear >= year)
    temp$year <- year ## add year column
    temp <- dplyr::select(temp, c(year, stateabb, ccode, statenme)) ## reorder
    isaf_ext <- rbind(isaf_ext, temp) ## append
}
```

# CINC Scores

```{R}
cinc <- readr::read_csv('../inst/extdata/NMC_5_0.csv')
cinc <- dplyr::select(cinc, -c(version))
isaf_ext <- dplyr::left_join(isaf_ext, cinc)
```

# Ideal Points

Second, let's merge the Bailey et al. ideal point measures.

```{R}
## import data
ideal <- haven::read_dta('../inst/extdata/voeten_un_votes/Idealpoints.dta')

## prep for merging by removing unwanted variables
ideal <- dplyr::select(ideal, c(ccode, year, idealpoint, 
    pctagreeus))
colnames(ideal) <- c("ccode", "year", "idealpoint", "pctagreeus")

## merge data
isaf_ext <- dplyr::left_join(isaf_ext, ideal)
```

## Ideal point distance from US

+ Standardize ideal point to be between 0 and 1. 

```{R}
## Min-max
range_01 <- function(x){
    (x - min(x, na.rm = TRUE))/(
    max(x, na.rm = TRUE) - min(x, na.rm = TRUE))
    }

isaf_ext$idealpoint <- range_01(isaf_ext$idealpoint)
```

+ Difference between each state and US

```{R}
isaf_ext$ideal_dist <- 0
for(i in 1:nrow(isaf_ext)){
    time <- as.numeric(isaf_ext[i,1]) ## what year
    temp <- dplyr::filter(isaf_ext, year == time) ## only that year data
    us <- dplyr::filter(temp, ccode == 2)$idealpoint
    isaf_ext[i,14] <- abs(isaf_ext[i,12] - us)
}
```

# Polity

We include the polity scores in order to test for whether democracies serve as more reliable partners.

```{R}
polity <- rio::import('../inst/extdata/p4v2017.xls')
polity <- dplyr::filter(polity, year >= 2001)
## variables to merge on and keep
polity <- dplyr::select(polity, c(ccode, year, democ, autoc, polity2))
## create 0,1 variable for democracy
polity$dem <- 0
for(i in 1:nrow(polity)){
    if(polity[i,3] >= 6){
    polity[i, 6] <- 1
    }
}
## merge
isaf_ext <- dplyr::left_join(isaf_ext, polity)
```

# Alliance Depth

Note -- there are only a handful of unique values for alliance depth estimates because the number of discrete U.S. alliance treaties during our dataset is not large. There are, however, a large number of countries in those alliances. 

```{R}
depth_dyad <- rio::import('../inst/extdata/benson_clinton/depth_scores.Rds')
```

```{R}
## Use the year, country-code and depth for merging
depth_sub <- dplyr::select(depth_dyad, c("year", "mem2", "atopid", "depth"))
colnames(depth_sub) <- c("year", "ccode", "atopid", "depth")
## Note, there do not appear to be repeat country-years.
isaf_ext <- dplyr::left_join(isaf_ext, depth_sub)
```

+ Standardize depth to be between 0 and 1. Afghanistan is the low outlier.

```{R}
# log all values once min is added -- deals with outlier well
isaf_ext$depth <- log(isaf_ext$depth + 2.141)
# recode -Inf to min value
isaf_ext$depth[is.infinite(isaf_ext$depth)] <- 0
```


# US ally variable

```{R}
## Create alliance (not including non-aggression pact) variable
isaf_ext$us_ally <- ifelse(
    !is.na(isaf_ext$depth), ## if there is a depth score
    1, 0 ## 0 if no depth score (no alliance), 1 if there is a depth score
)
```

# Potential gain variable:

```{R}
## If no alliance, then depth goes to 0
isaf_ext$depth[isaf_ext$us_ally == 0] <- 0
## need to recode to reflect closer to US is better
## Just flip by substracting from 1. So 0 (no dist) becomes 1
## and 0.9 (very different) becomes 0.1 (low alignment)
isaf_ext$pot_gain <- (1 - isaf_ext$ideal_dist) - isaf_ext$depth
## drop US
isaf_ext$pot_gain[isaf_ext$ccode == 2] <- NA
```

# DV

Let's import the iiss data about troop levels in Afghanistan.

```{R}
iiss_troops <- rio::import('../data/iiss_afghan.rds')
iiss_troops <- dplyr::arrange(iiss_troops, year)
```

There are some repeats because of data listed in separate places. Clean and combine.

```{R}
dups <- which(duplicated(iiss_troops[,1:2]) == TRUE)

for(i in 1:length(dups)){
    ## troops_afghan_total
    iiss_troops[dups[i] - 1,5] <- iiss_troops[dups[i] - 1,5] + 
        iiss_troops[dups[i],5]
    ## troops_afghan_oef
    iiss_troops[dups[i] - 1,6] <- iiss_troops[dups[i] - 1,6] + 
        iiss_troops[dups[i],6]
    ## troops_afghan_isaf
    iiss_troops[dups[i] - 1,7] <- iiss_troops[dups[i] - 1,7] + 
        iiss_troops[dups[i],7]
    ## troops_afghan_unama
    iiss_troops[dups[i] - 1,8] <- iiss_troops[dups[i] - 1,8] + 
        iiss_troops[dups[i],8]
    ## troops_afghan_ratio -- recalculate
    iiss_troops[dups[i]-1,9] <- iiss_troops[dups[i]-1,5] / 
        iiss_troops[dups[i]-1,4] 
}

## remove dups
iiss_troops <- iiss_troops[!duplicated(iiss_troops[,1:2]),]
```

# Controls

+ Distance from Afghanistan, GDP, GDPPC, and casualties.

```{R}
## Distance
distance <- readr::read_csv('../inst/extdata/distance.csv')
## Afghanistan = 700
afgh_dist <- dplyr::filter(distance, ccode1 == 700 | ccode2 == 700)
## Not sure how to do this elegantly... Just going with brute force.
afgh_dist2 <- dplyr::tibble(
    ccode = as.integer(),
    distance = as.integer()
)
for(i in 1:nrow(afgh_dist)){
    ## Ccode -- get non-Afghanistan country
    if(afgh_dist[i,2] == 700){ 
        afgh_dist2[i,1] <- afgh_dist[i,1] 
    }else(
        afgh_dist2[i,1] <- afgh_dist[i,2]
    )
    ## Distance
    afgh_dist2[i,2] <- afgh_dist[i,4]
}
## which are unique?
unique_countries <- afgh_dist2[!duplicated(afgh_dist2[,c('ccode')]),]

## Merge with isaf
isaf_ext <- dplyr::left_join(isaf_ext, unique_countries)
```

+ GDP and GDPPC

```{R}
gdp <- wbstats::wb(indicator = c("NY.GDP.MKTP.CD"))
gdp <- dplyr::select(gdp, c(iso3c, date, value))
colnames(gdp) <- c("stateabb", "year", "gdp")
gdp$year <- as.numeric(gdp$year)
isaf_ext <- dplyr::left_join(isaf_ext, gdp)

gdppc <- wbstats::wb(indicator = c("NY.GDP.PCAP.CD"))
gdppc <- dplyr::select(gdppc, c(iso3c, date, value))
colnames(gdppc) <- c("stateabb", "year", "gdppc")
gdppc$year <- as.numeric(gdppc$year)
isaf_ext <- dplyr::left_join(isaf_ext, gdppc)
```

+ Casualties

```{R}
casualties <- rio::import('../data/afghan_casualties.csv')
## Tidy
## remove total column
casualties <- casualties[,-31]
casualties <- tidyr::gather(data = casualties, "country", "casualties", 2:30)
colnames(casualties) <- c("year", "statenme", "casualties")
## fix naming inconsistencies for emrging
casualties[501:522,2] <- "United States of America"
## Merge
isaf_ext <- dplyr::left_join(isaf_ext, casualties)
## If NA for casualties, then have equal 0
isaf_ext$casualties <- ifelse(is.na(isaf_ext$casualties), 0,    
    isaf_ext$casualties)
```

## Merge

+ Now we can merge everything together.

```{R}
iiss_troops$ccode <- as.integer(iiss_troops$ccode)
final_data <- dplyr::left_join(iiss_troops, isaf_ext)
```

+ And drop Northern Cyprus, because not internationally recognized and their troops were not available to Cyprus.

```{R}
final_data <- dplyr::filter(final_data, 
    country != "Turkish Republic of Northern Cyprus")
```

## Add NATO Placemats

```{R}
nato <- readr::read_csv("../data/nato_placemats.csv")
nato$count_total <- as.integer(nato$count_total)
## remove unecessary columns
nato$country <- as.character(nato$country)

colnames(nato) <- c("year", "country", "nato_troop_reported")

## there are multiple tallies per year, so let's make a new df only with max for year
nato_contr <- dplyr::tibble(
    year = as.integer(),
    country = as.character(),
    troops = as.integer()
)

countries <- unique(nato$country)
for(i in 1:length(countries)){
    for(j in 1:8){
        nato_contr[i*j, 1] <- j + 2006 ## year
        nato_contr[i*j, 2] <- countries[i] ## country
        temp <- dplyr::filter(nato, year == j + 2006 & country == countries[i]) 
        nato_contr[i*j, 3] <- max(temp$nato_troop_reported, na.rm = TRUE)
    }
}
## -inf returned when no value for that country-year
nato_contr$troops[is.infinite(nato_contr$troops)] <- NA
## remove NA rows
nato_contr <- nato_contr[complete.cases(nato_contr),]

## join
final_data <- dplyr::left_join(final_data, nato_contr)
```

## Troop contributions, Conditional on NATO data

```{R}
final_data$troops_afghan_final <- 0
for(i in 1:nrow(final_data)){
    if(is.na(final_data[i,34])){
        final_data[i,35] <- final_data[i,5]
    }else{
        final_data[i,35] <- final_data[i,34]
    }
}

final_data$troops_afghan_final_ratio <- final_data$troops_afghan_final/     
    final_data$troops_total
```


# Finalize

+ Last, let's clean everything up.

```{R}
## Remove redundant id var
final_data <- dplyr::select(final_data, -c(statenme)) 
## Reorder so that stateabb is not in the middle of the dataset
final_data <- dplyr::select(final_data, c(year, country, stateabb, 
    dplyr::everything()))
```

# Save files

```{R}
iiss_rhs <- dplyr::select(final_data, -c(troops_total, troops_afghan_total,
    troops_afghan_oef, troops_afghan_isaf,
    troops_afghan_unama, troops_afghan_ratio,
    troops_afghan_final,
    troops_afghan_final_ratio))

## right hand side data
rio::export(iiss_rhs, '../data/iiss_afghan_rhs.rds')
## merged complete data
rio::export(final_data, '../data/iiss_afghan_complete.rds')
```
