==> docs/05_Model.Rmd <==
---
title: "05 Model"
author: "J Andres Gannon and Daniel Kent"
output:
  html_document:
    theme: flatly
    code_download: true
    code_folding: hide
    toc: yes
    number_sections: yes
    toc_float:
      collapsed: yes
  html_notebook:
    toc: yes
    toc_float: yes
editor_options:
  chunk_output_type: inline
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(warning = FALSE, error = FALSE)
library(magrittr)
library(ggplot2)
```

# Data loading

```{R}
isaf <- rio::import('../data/iiss_afghan_complete.rds')
## Remove afghanistan because it is not a realistic self-contributor
## Remove US because paper is about relations with the US
isaf <- dplyr::filter(isaf, stateabb != "AFG" & stateabb != "USA")
```

+ Iceland -- denominator is technically 0, but they sent some police officers. We drop for parsimony.

```{R}
isaf <- dplyr::filter(isaf, stateabb != "ICE")
```

# Description and visualization

+ Summary stats

```{R}
## Remove id variables
descrip <- dplyr::select(isaf, -c(year, country, ccode, stateabb))
## Descriptive grid
GGally::ggpairs(dplyr::select(descrip, 
  -c(troops_total, troops_afghan_total, troops_afghan_oef, troops_afghan_isaf,
  troops_afghan_unama, milex, milper, irst, pec, tpop, upop, atopid, pot_gain,
  gdp, gdppc, casualties, distance))) + ggplot2::theme_bw()
```

+ Ok, so looks good. Some important points to consider. Notice that, by construction, `idealpoint` and `pctagreeus` are strongly associated with `pot_gain`. 

## Distribution of troops

+ Custom font

```{R}
## Custom font
library(showtext)
## Slabo 27px is close enough to bitstream
## Need to download slabo 27 from google fonts and install
font_add_google(name = "Slabo 27px")
font_add("Slabo", "Slabo27px-Regular.ttf")
showtext_auto() ## automatically use showtext for new devices
myfont1 <- "Slabo"
```

+ Boxplot by year

```{R}
library(ggplot2)
troop_year <- ggplot(dplyr::filter(isaf, troops_afghan_total > 0)) +
  geom_boxplot(aes(x = year, y = troops_afghan_total, 
    group = year), size = 0.5, color = "#004570") + 
    theme_bw() +
    theme(text = element_text(size = 16, family = myfont1),
        axis.text.x = element_text(colour="black"), 
        axis.text.y = element_text(colour="black")) + 
    scale_x_discrete(name ="Year", 
                    limits=c(2001:2013)) + 
    labs(title = "Troop Contributions to ISAF by Year",
        subtitle = "2001 - 2013",
        y = "Troops")

ggsave(troop_year, file = "../paper/figures/troop_year.pdf",
      width = 8, height = 5, dpi = 1000)
```

+ Troop percent

```{R}
troop_dist <- ggplot(isaf, aes(x = troops_total)) + 
  geom_histogram() + #binwidth = 0.01) + 
  theme_bw() +
  theme(text = element_text(size = 14, family = myfont1),
        axis.text.x = element_text(colour="black"), 
        axis.text.y = element_text(colour="black")) + 
  labs(title = "Troop Contributions to ISAF",
    subtitle = "2001 - 2013",
    x = "Percent of Total Troops",
    y = "Frequency")

ggsave(troop_dist, file = "../paper/figures/troop_dist.pdf",
      width = 5, height = 5, dpi = 1000)
```

# Models

```{R}
m1 <- glm(
  troops_afghan_final_ratio ~ pot_gain, 
  data = isaf
  ) 
summary(m1)
## Just outside of significance


## Keeping out gdp and gdppc because they drop so many observations
## Drop pctagreeus cuz used for indicator construction
m2 <- glm(
  troops_afghan_final_ratio ~ pot_gain + pctagreeus + dem + casualties + us_ally + cinc,
  data = isaf
  )
summary(m2)


## And let's control for gdp and gdppc and keep pctagreeus -- though loses obs
m3 <- glm(
  troops_afghan_final_ratio ~ pot_gain + pctagreeus + dem + casualties + us_ally + cinc + log(gdp) + log(gdppc) + distance,
  data = isaf
  )
summary(m3)

## fixed effects? 
## note distance coefficient not estimated, because it has 
## no within-unit variation
## gdp and gdppc dropped because lack of observations 
## worsens fixed effects bias-variance tradeoff
m4 <- plm::plm(
  troops_afghan_final_ratio ~ pot_gain + 
  dem + casualties + us_ally + cinc,
  model = "within",
  index = c("ccode", "year"),
  data = isaf)
summary(m4)

texreg::texreg(
  list(m1, m2, m3, m4),
  custom.model.names = c(
    "Baseline",
    "Political Controls",
    "All Controls",
    "Fixed Effects"
  ),
  custom.coef.names = c(
    "Intercept", "Potential Gain", "UN Voting", "Democracy", 
    "Casualties", "U.S. Ally", "CINC Score", "Log GDP",
    "Log GDPPC", "Distance"),
  digits = 3,
  ci.force = TRUE,
  ci.force.level = 0.95,
  file = "../paper/table.tex"
)


## Latex table -- https://github.com/vincentarelbundock/gtsummary
#models <- list()
#models[['Model 1: Baseline']] <- m1
#models[['Model 2: Political Controls']] <- m2
#models[['Model 3: All Controls']] <- m3
#models[['Model 4: Fixed Effects']] <- m4
#gtsummary::gtsummary(
#  models, 
#  statistic = 'conf.int', 
#  conf_level = .95,
#  file = "../paper/table.tex") 


## And let's build figures for contributions -- use code from mershon
library(dotwhisker)
library(dplyr)
dw1 <- dwplot(
  list(m1, m2, m3, m4), dot_args = list(size = 2.5),
    whisker_args = list(size = 1.5), 
    color = "black") %>%
    relabel_predictors(c(
        pot_gain = "Potential Gain With U.S.",
        idealpoint = "Ideal Point Similarity With U.S.",
        dem = "Democracy",
        casualties = "Casualties",
        us_ally = "U.S. Ally", 
        cinc = "CINC",
        gdp = "Log GDP",
        gdppc = "Log GDPPC",
        distance = "Distance"
        )) +
    geom_vline(xintercept = 0) +
    ## labs(x = "Perceived Threat (0 to 4 Scale)") +
    ggtitle("Coefficient Estimates") +
        theme(plot.title = element_text(face="bold", hjust = -0.0),
        legend.justification=c(0, 0), legend.position=c(0, 0),
        legend.title = element_blank()) +
    scale_colour_discrete(breaks=c(
      "Model 1", "Model 2", "Model 3", "Model 4"
      ),
        labels = c(
        "Potential Gain Only", 
        "Political Controls", 
        "All Controls", 
        "Fixed Effects")) +
    labs(x = "Percent of Total Troops Provided") + 
    theme_bw() +
    theme(text = element_text(size = 24),
        axis.text.x = element_text(colour="black"), 
        axis.text.y = element_text(colour="black"),
        legend.position = c(0.8, 0.01),
        legend.justification = c(0, 0), 
        legend.background = element_rect(colour="grey80"),
        legend.title = element_blank())

## Save the figure
ggsave(dw1, file = "../paper/figures/coefs.pdf",
      width = 19, height = 8, dpi = 600)
```


## Predicted contributions

+ Predicted contributions? Let's hold all variables at their central value (mean or median, depending on the distribution) and then look at the predicted troop ratio for each value of potential gain.

```{R}
## Simulated Data
us_ally0 <- with(
    isaf,
    dplyr::tibble(
      pot_gain = seq(
        0, #min(pot_gain, na.rm = TRUE),
        1, #max(pot_gain, na.rm = TRUE),
        length = 101
        ),
      pctagreeus = mean(pctagreeus, na.rm = TRUE),
      dem = 1,
      casualties = mean(casualties),
      us_ally = 1,
      cinc = median(cinc, na.rm = TRUE),
      gdp = median(gdp, na.rm = TRUE),
      gdppc = median(gdppc, na.rm = TRUE),
      distance = median(distance, na.rm = TRUE)
      )
    )

## Predicted contributions
## useful walkthrough: https://www.fromthebottomoftheheap.net/2017/05/01/glm-prediction-intervals-i/

us_ally0$pred_contr <- predict(m3, newdata = us_ally0)

pred_int <- predict(m3, newdata = us_ally0, se.fit = TRUE)

pd <- transform(pred_int,
  Contributions = fit,
  Upper = fit + (1.96 * se.fit),
  Lower = fit - (1.96 * se.fit),
  Gain = seq(from = 0, to = 1, by = 0.01)
  )

## Visualize
pred_fig <- ggplot(pd) +
    geom_ribbon(aes(x = Gain, ymin = Lower * 111976, ymax = Upper * 111976),
                fill = "steelblue2",
                alpha = 0.3, inherit.aes = FALSE) +
    geom_line(aes(x = Gain, y = Contributions * 111976), size = 1.25) +
    geom_line(aes(x = Gain, y = Lower * 111976), size = 1) +
    geom_line(aes(x = Gain, y = Upper * 111976), size = 1) +
    labs(
      x = "Potential Alliance Gain (Maximum Value at 1)",
      y = "Predicted Contributed Troops",
      title = "Predicted Contributions by Potential Alliance Gain With the United States") +
    ylim(c(500, 1500)) +
    hrbrthemes::theme_ipsum() +
    theme(
      text = element_text(size = 12),
      axis.text.x = element_text(color = "black", size = 12),
      axis.text.y = element_text(color = "black", size = 12),
      axis.title.x = element_text(color = "black", size = 12),
      axis.title.y = element_text(color = "black", size = 12),
      title = element_text(size = 14)
      )
      ## Median troop number is 26,500 -- discuss what this means in caption

tikzDevice::tikz(width = 8, height = 5, 
  file = "../paper/figures/pred_contr.tex")
pred_fig
dev.off()


## Save the figure
ggsave(pred_fig, file = "../paper/figures/pred_contr.pdf",
      width = 8, height = 5, dpi = 1000)
```

+ Additional predictions figure, take average country and show the predicted increase if they go from current place to maximum or decrease to 0.

```{R}
us_ally1 <- with(
    isaf,
    dplyr::tibble(
      pot_gain = 0.2,
      pctagreeus = mean(pctagreeus, na.rm = TRUE),
      dem = 1,
      casualties = mean(casualties),
      us_ally = 1,
      cinc = median(cinc, na.rm = TRUE),
      gdp = median(gdp, na.rm = TRUE),
      gdppc = median(gdppc, na.rm = TRUE),
      distance = median(distance, na.rm = TRUE)
      )
    )

current_pred <- predict(m3, newdata = us_ally1)

# SD for pot_gain is 0.338, so will range by 0.35 from average ~ 0.2
# Now make counterfactual df with s.e.
us_ally2 <- with(
    isaf,
    dplyr::tibble(
      pot_gain = seq(
        0.2 - 0.35, # near min observed -- makes symmetric around avg val
        0.2 + 0.35, # near max observed
        length = 25
        ),
      pctagreeus = mean(pctagreeus, na.rm = TRUE),
      dem = 1,
      casualties = mean(casualties),
      us_ally = 1,
      cinc = median(cinc, na.rm = TRUE),
      gdp = median(gdp, na.rm = TRUE),
      gdppc = median(gdppc, na.rm = TRUE),
      distance = median(distance, na.rm = TRUE)
      )
    )
  

pred_int1 <- predict(m3, newdata = us_ally2, se.fit = TRUE)

pd2 <- transform(pred_int1,
  Contributions_alt = fit,
  Upper = fit + (1.96 * se.fit),
  Lower = fit - (1.96 * se.fit),
  pot_gain = us_ally2$pot_gain
  ) %>%
  select(c(Contributions_alt, Upper, Lower, pot_gain))

pd2 <- pd2 %>%
  mutate(
    diff_pot = pot_gain - 0.2,
    diff_cont = Contributions_alt - current_pred,
    diff_lower = Lower - current_pred,
    diff_upper = Upper - current_pred
  )

pred_fig2 <- ggplot(pd2) +
  geom_line(aes(x = diff_pot, y = diff_cont * 111976), size = 1.25) + 
  geom_ribbon(
    aes(x = diff_pot, ymin = diff_lower * 111976, ymax = diff_upper * 111976),
    fill = "steelblue2",
    alpha = 0.3, inherit.aes = FALSE
    ) + 
    geom_line(aes(x = diff_pot, y = diff_lower * 111976), size = 1) +
    geom_line(aes(x = diff_pot, y = diff_upper * 111976), size = 1) +
  hrbrthemes::theme_ipsum() + 
  labs(
    x = "Change in Unrealized Alliance Potential",
    y = "Predicted Change in Contributed Troops",
    title = "Predicted Change in Contributed Troops \nBy Change in Unrealized Alliance Potential With the United States"
  ) +
  scale_x_continuous(breaks = seq(-0.35, 0.35, by = 0.1)) +
  scale_y_continuous(breaks = seq(-300, 400, by = 150)) +
    theme(
      text = element_text(size = 12),
      axis.text.x = element_text(color = "black", size = 12),
      axis.text.y = element_text(color = "black", size = 12),
      axis.title.x = element_text(color = "black", size = 12),
      axis.title.y = element_text(color = "black", size = 12),
      title = element_text(size = 14)
      )

tikzDevice::tikz(width = 8, height = 5, 
  file = "../paper/figures/pred_contr2.tex")
pred_fig2
dev.off()
```

## Robustness Checks -- Other DVs

Troops as a percent of GDP

```{R}
isaf$troops_gdp <- isaf$troops_afghan_total/isaf$gdp

m1b <- glm(
  troops_gdp ~ pot_gain, 
  data = isaf
  ) 
summary(m1b)
## Just outside of significance


## Keeping out gdp and gdppc because correlated with DV
m2b <- glm(
  troops_gdp ~ pot_gain + pctagreeus + dem + casualties + us_ally + cinc,
  data = isaf
  )
summary(m2b)
```

Troops as a percent of population

```{R}
isaf$troops_pop <- isaf$troops_afghan_total/isaf$tpop

m1c <- glm(
  troops_pop ~ pot_gain, 
  data = isaf
  ) 
summary(m1c)
## Just outside of significance


## Keeping out gdp and gdppc because they drop so many observations
## Drop pctagreeus cuz used for indicator construction
m2c <- glm(
  troops_pop ~ pot_gain + pctagreeus + dem + casualties + us_ally + cinc,
  data = isaf
  )
summary(m2c)


## And let's control for gdp and gdppc and keep pctagreeus -- though loses obs
m3c <- glm(
  troops_pop ~ pot_gain + pctagreeus + dem + casualties + us_ally + cinc + log(gdp) + log(gdppc) + distance,
  data = isaf
  )
summary(m3c)

## fixed effects? 
## note distance coefficient not estimated, because it has 
## no within-unit variation
## gdp and gdppc dropped because lack of observations 
## worsens fixed effects bias-variance tradeoff
m4c <- plm::plm(
  troops_pop ~ pot_gain + 
  dem + casualties + us_ally + cinc,
  model = "within",
  index = c("ccode", "year"),
  data = isaf)
summary(m4)
```

## Map location

+ Set up world map data:

```{R}
isaf <- rio::import('../data/iiss_afghan_complete.rds')
isaf <- dplyr::filter(isaf, stateabb != "ICE") ## Iceland
library(ggplot2)
WorldData <- map_data('world')
isaf_map <- dplyr::select(isaf, c("year", "country", 
  "troops_afghan_ratio", "troops_afghan_total"))
isaf_map$country <- as.character(isaf_map$country)


## Labels -- Make sure the same for both datasets
for(i in 1:nrow(isaf_map)){
  ## Korea, Republic
  if(isaf_map[i,2] == "Korea, Republic of"){
    isaf_map[i,2] <- "South Korea"
  }
  ## Macedonia
  if(isaf_map[i,2] == "Macedonia, Former Yugoslav Republic of"){
    isaf_map[i,2] <- "Macedonia"
  }
  ## United Kingdom
  if(isaf_map[i,2] == "United Kingdom"){
    isaf_map[i,2] <- "UK"
  }
  ## USA
  if(isaf_map[i,2] == "United States"){
    isaf_map[i,2] <- "USA"
  }
}

isaf_map2 <- dplyr::tibble(
  region = as.character(),
  Contribution = as.numeric(),
  Troops = as.integer()
)

for(i in 1:length(unique(isaf_map$country))){
  ## Subset by unique obs
  temp <- dplyr::filter(
    isaf_map, country == unique(isaf_map$country)[i])
  isaf_map2[i,1] <- unique(isaf_map$country)[i]
  ## Most ever contributed
  isaf_map2[i,2] <- max(temp$troops_afghan_ratio)
  isaf_map2[i,3] <- max(temp$troops_afghan_total)
}

countries <- unique(isaf_map2$region)
countries2 <- unique(WorldData$region)
missing_countries <- setdiff(countries2, countries)

for(i in 1:length(missing_countries)){
  #row <- i + 173
  temp <- dplyr::tibble(
    region = as.character(),
    Contribution = as.numeric(),
    Troops = as.integer() 
  )
  temp[1,1] <- missing_countries[i]
  temp[1,2] <- 0
  temp[1,3] <- 0

  ## append
  isaf_map2 <- rbind(isaf_map2, temp)
}

## Theme_map() https://github.com/jrnold/ggthemes/blob/master/R/theme-map.R
library(dplyr)
theme_map <- function(base_size = 9, base_family = "") {
  theme_bw(base_size = base_size, base_family = base_family) %+replace%
    theme(axis.line = element_blank(),
          axis.text = element_blank(),
          axis.ticks = element_blank(),
          axis.title = element_blank(),
          panel.background = element_blank(),
          panel.border = element_blank(),
          panel.grid = element_blank(),
          panel.spacing = unit(0, "lines"),
          plot.background = element_blank(),
          legend.justification = c(0, 0)#,
          #legend.position = c(1,1)
          )
}

isaf_map2$perc_total <- isaf_map2$Contribution * 100

isaf_map2$perc_isaf <- isaf_map2$Troops/sum(isaf_map2$Troops)
isaf_map2$perc_isaf <- isaf_map2$perc_isaf * 100

contr_map2 <- ggplot() +
  geom_map(data = WorldData, map = WorldData,
    aes(x = long, y = lat, group = group, map_id = region),
    fill = "white", size = 0.35) +
  geom_map(data = isaf_map2, map = WorldData,
    aes(fill = perc_isaf, map_id = region)) +
  borders("world", size = 0.425, colour = "black") +
  scale_fill_gradientn(
    colours = c(
      "white", "#adccff", "#004570"), 
    values = c(0, 0.025,1 )) +
    theme_map() + 
    labs(title = "Percent of Total ISAF Coalition Contributed",
      fill = "Percent") 

## Save the figure
ggsave(contr_map2, file = "../paper/figures/contr_map_isaf.pdf",
      width = 8, height = 4, dpi = 600)
```
